requirements.txt
streamlit
supabase
pandasstreamlit
supabase
st-supabase-connection 